Removed cached versions of a bunch of files that should be ignored
They should not have been added to the repository, but they were so this
is correcting that. These files all start with a line to the effect of
"Don't edit this because it's automatically going to be replaced next
time you build" So if you think you need these files, go looking for the
source that generates them instead.