Added a python based automated build for the extension
I wanted to make a cross platform build script, and I noticed we already
require python to point to version 2.7.x so I figured I could use this
to make a quick and dirty cross platform build script for the extension
based test and app environment. This way we can all use it and
collaborate on it and I never have to help someone debug somehting where
they just forgot a step.