more automated build tools
install.py was a slightly confusing name for reasons I don't want to go
into so I renamed it install_extension.py which is enough to get the new
app in developer mode
install_app.py lets you build the app with a new version number so it
can be uploaded from the release directory. (I don't see a zip file
there, so maybe grunt compress doesn't work properly on windows? I'm not
sure)
update_versionNumber.py is a helper function to update the version
number in manifest-src/manifest-head.json