Did some work to add the unbuilt test to the repository
It's not done yet, but the generator files can now be tested in place by
opening static/demos/code/index.html
Can also run the shortcut in the main directory.
Some files may be duplicated. It's worth updating the path to point to
the same files as are used in the build when these are encountered.