fixed bug with procedure names that contain a space
this used to cause an error because blocky automatically replaces space
with underscore in a var or proc name when it is added to the Blockly
variable name DB, but since we were picking up the name from the block,
not the variableDB, the name did not not match the block name. causing
an unknown variable error.