handle certain unresolved variable refs
Sometimes a variable is incompletely defined (like create list with 3
elements but don’t specify the element length). In this case memory
cannot be allocated to the variable. But if the variable is never used
we don’t care about it and should not break assembly generation. So we
don’t generate code to create that variable.